requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LLM and AI
sentence-transformers>=2.3.0
httpx>=0.26.0

# Vector Database
pinecone-client>=3.0.0

# Frontend
streamlit>=1.30.0

# Utilities
numpy>=1.24.0
python-multipart>=0.0.6

# Optional: Uncomment if you want to use langchain
# langchain>=0.1.0
# langchain-community>=0.0.16
